Remove unnecessary Register Interface.
Register Interface provided a clean abstraction over the
`RegisterPairType`, however, it grossly simplified the interface down to a
point where it would be impossible to fix without either
over-complicating the abstractions or over-simplifing
`RegisterPairType`. Because of this, the best approach seems to be
removing the abstractions and working directly with `RegisterPairType`.