Actually implement LD [R8], IMM
The previous implementation was both broken and not good. We removed the
overly-complicated function pointers and moved some decoding logic into
the instructions themselves. While this may be a bit code-duplicatey,
it's much simpler to work with and doesn't have the bugs of the previous
version.

All test cases pass, so that's good. Don't forget the flags register in
the future.