Add .clang-format.
For auto-formatting I've supplied a custom configuration based on the
WebKit linting rules, modified to fix up some of my personal
preferences. With this configuration, the code can now be properly
formatted automagically. All future commits should be certain to use
these formatting rules so the codebase is cleaner and consistent.