Format code files.
Having added the formatter rules, I've now run the formatter on all C++
files in the project so they are now all formatted correctly. Future
commits _should_, ideally, be formatted before being pushed.

A lot of these changes are being made to ease the transition from Qt
Creator, which is a mostly fine C++ editor, to other editors which are a
little less annoying to use, such as VS Codium, and that have easier
integration with autoformatters and linters.

My _hope_ with these changes is that I will be able to move to a
different editor for working on this project, as QT Creator is fairly
heavy, takes a while to load, and has many bugs when it comes to
auto-completion, largely some popup messages will begin flicking wildly
until I do something specific to make them go away (this issue has also
been noted on multiple versions of Plasma 5, and is likely directly
related to QT5, the version I use for my applications).

In essence: this change will also allow enforcing of coding standards,
with a specific example in the form of the configuration, for any PRs
made to the project.