Implement VS Codium configurations.
These configurations allow any users who want to use one of the VSCode
family editors, in my case Codium, to easily get started with
automation. Included are two files: `launch.json` for setting up
automated debugging, with Codium integration for GDB, and `tasks.json`
for automating the build step before launching GDB. This ensure anyone
using a VSCode-based editor to quickly get into an easy debugging
environment that's guaranteed to always be built ahead of time, so they
don't have to remember to rerun CMAKE before debugging. It also ensures
the configuration is set to Debug, so there will be debug symbols
available to the debugger.

Coming in the future: a way to quickly setup profiles for debug and
release, a way to enable and disable testing, and more configuration
options within the CMAKE system.