Fix half-carry logic for subtraction.
If I'm correct, this should now be the ACTUAL half-carry logic.
I'm not 100% certain it _is_ correct, but it makes more sense. Tests
have been adjusted to account for the changes. However, this change
causes almost all subtraction-based instructions to now set the
half-carry. This may be the correct behaviour? It's very unclear. I
don't know if the Z80 takes the absolute value of a number in subtract
mode or if it calculates by the binary representation.

This is important, as take the following example:
> 1 - 1

If it takes the binary representation and adds the two numbers, the
half-carry will be set, because 0b11111111 > 0b00000001. If it encodes
the subtract separately from the value, then the half-carry should not
be set, as 0b00000001 == 0b00000001, and no carry would occur.