Remove `.vscode` from the project.
Unlike the previous two, this is actually not being done because of an
accidental inclusion due to using the wrong Git flag (that is what
happened). These files are being removed solely because they don't work.
I was playing around with using VSCodium instead of QT Creator, but ran
into some problems. These files, in their current state, are _not
usable_! I would recommend anyone using VSCode/Codium/Code OSS remove
these files and create their own! At least until I can revisit that
challenge and try again, at which point the files will be added back
once they are known to work.

I am not inherently opposed to including editor files in a source tree,
it does aid in others using the same tools to quickly get up-and-running
in a known-working environment. However, as mentioned, these would be
more problematic and lead to issues immediately after opening the
editor. It's not on others to fix issues I created, so for simplicity
these are being removed.