Implement ld [r], [r] instructions.
This was... A massive pain. There's a lot to unpack, 7 new test cases,
and a lot of suffering. Absolutely the worst testing functions I've
written so far, and part of that was simply how tedious it was.

The code is not very clean, but it works, the cases pass, and there are
some things I can clean up in the future that should make things much
easier. The worst part is that I'm seriously considering dropping Qt6 in
favour of another GUI framework, which means rewriting all of the tests
to use a different suite. That will be fun... For now, though, this
works. Also, amzingly, the actual logic itself was only about 15 lines
of code. So the problem was simply testing every permutation...
Separating register A from the rest definitely makes things interesting.
Having `(HL)` treated as a single register is also fun. Imagine needing
specific testing functions for `ld (hl), a` and `ld a, (hl)`...