Fix testing crash on tst_ld_r_addr_hl.
This was a bit of a longer change that took _days_ of figuring things
out. Ultimately, it turned out Valgrind was the answer I was looking
for, and fixing all of the dangling pointers and code smell instantly
made things so much easier!

First: no more C-style pointer, the CPU is a `unique_ptr<>`.
Second: `init` and `cleanup` are much simpler. The entire test suite
uses the same `cpu` and `bus`. At the end of each test, the `bus` is
reset to initial state, but the CPU is left in the previous state. This
behaviour _may_ change in the future once testing is made more robust
and we ensure state only changes where necessary.

With all of these changes, I've reduced the number of errors reported by
Valgrind to 1, which at this time I believe to be an issue with GLIBC
and not the codebase. This _could_ be a misunderstanding in how
`std::map<>` works, but the standards doc states that my code _should_
be fine. Unless I'm wrong, but the issue doesn't seem to impact the
testing currently so I've chosen to ignore it.

All tests now pass once again, and I've resolved the weird behaviour
with `qInfo()` that caused the massive headache.