Add beginning support for port-mapped devices.
This is a fairly large one, mostly because it touched a good chunk of
the codebase. For starters, the `MmioDeviceManager` now supports
port-mapped devices, denoted by the `read8()` and `write8()` methods
taking a new argument, the `iorq` argument. On physical hardware, this
is a pin that determines if the CPU is calling for port-mapped or
memory-mapped addresses.

Because of this change, anywhere in the codebase that called out to
`read8()` or `write8()` has been rewritten. This is where most of the
changes were made.

The `MmioDeviceManager` has had a new method added to allow adding
port-mapped devices, as well as helper methods to differentiate between
port-mapped and memory-mapped reads and writes. Of course, with such a
major change this also means the entire testing suite for
`MmioDeviceManager` has been updated, with each method (private and
public) extensively tested for proper behaviour.

In hindsight, I should have given the `read8()` and `write8()` methods a
default value for `iorq`, to maintain compatible ABI... I forgot that
was a thing, too much writing Rust lately.

Final note: `read16()` and `write16()` have been left untouched, as
port-mapped devices only use 8-bits of the address space. A port-mapped
device cannot send two bytes back-to-back, so if such a behaviour is
necessary it would require custom Assembly logic in the emulation.