Remove in-code TODOs.
A small cleanup, future TODOs will be logged in an issue tracker.
Currently I am using the issue tracker in a locally-hosted, private
Forgejo instance, but once the project is ready for use the issues will
be migrated to a public source, likely GitHub if I don't move to
Codeberg.

If anyone _does_ read this, know that as of right now, any commits that
reference an issue number refer specifically to the private issue
tracker. I will try to specifically indicate whether a commit refers to
a local issue, and I'm currently uncertain what will happen when moving
to GitHub.

The local issue tracker is strictly for early development changes, as a
large chunk of this project is currently documented in a Joplin Notebook
that only I have access to. All of this information is planned to be
moved into a Wiki once the project is usable (a lot of infrastructure
exists locally to simplify the development process). The local tracker
just eases keeping up with goals that I want to work towards, as well as
getting comfortable with self-hosted solutions.