Remove parameter from instruction methods.
Closes #3 on the local, private issue tracker. The Z80BaseCpu class
already contains a reference to the bus, which means there's no reason
to pass a redundant reference to the bus to all instruction methods.
Those methods now explicitly use the class' internal reference.

Since I prefer explicit `this` over implicit, this also means every
method that uses the bus has been modified to prefix `this->` before the
call to the bus.