Fix logic bug in `ADD A, r` for carry flag.
Fixes #6 in the private, local issue tracker, I think. The carry flag
should now be properly set in the addition instruction. I've compared
the behaviour with another emulator and the flag is set.

In short, the carry flag would never be set because the code was
checking if the signed integer is greater than 255, but this check has
been switched to unsigned as otherwise it could never trigger. Tests
have been updated to reflect the correct behaviour of the carry flag.