Instructions Consistently Return UINT8_MAX.
Fixes #12 in the local, private issue tracker. Instructions were
inconsistently returning 255 _or_ UINT8_MAX, now they all return
UINT8_MAX. While the two values should always be identical, this makes
it clearer that we are returning the maximum number of cycles possible
rather than an arbitrary value.

This is subject to change in the future, as it's still unclear _why_ the
code returns max 8-bit integer. To explain here: returning 255 is done
to indicate an error state that should be entirely impossible to
trigger. Neither C nor C++ have any way to indicate a type that is
smaller than one byte, and we are testing on values generated by parts
of a byte. In this case, the switch statement should never be given a
value larger than 7. But otherwise there is a warning generated that not
all paths are accessible and could cause issues.