Move instruction decoding logic to method.
- Fixes #9 in the local, private issue tracker.

Very minor refactor. With this change, I can move on to resolving #10 as
well as run tests explcitly for the decoding logic without having to
call the `decode` method. For full context, the original issue proposed
moving the method to a helper function declared outside the class.
However, that would require making the `DecodedInstruction` struct
external to the class, and it has no purpose outside this specific
logic.